verifier/requirements.txt
google-genai>=1.0
near-api-py>=0.1.0
